+++ web/src/features/fireWeather/components/SidePanel.tsx
@@ -54,7 +54,10 @@ const SidePanel = (props: Props) => {
             <ToggleButton value={SidePanelEnum.Tables}>Tables</ToggleButton>
             <ToggleButton value={SidePanelEnum.Graphs}>Graphs</ToggleButton>
             {props.stationCodes.length > 1 && (
-              <ToggleButton value={SidePanelEnum.Comparison}>
+              <ToggleButton
+                value={SidePanelEnum.Comparison}
+                data-testid="station-comparison-button"
+              >
                 Station comparison
               </ToggleButton>
             )}
